Refactor: Split Index page into components
Split the Index page into smaller, reusable components to improve maintainability and readability.
