+++ src/hooks/useAnimateOnScroll.tsx
@@ -1,0 +1,27 @@
+
+import { useEffect } from 'react';
+
+const useAnimateOnScroll = () => {
+  useEffect(() => {
+    const handleScroll = () => {
+      const elements = document.querySelectorAll('.animate-on-scroll');
+      
+      elements.forEach(element => {
+        const rect = element.getBoundingClientRect();
+        const windowHeight = window.innerHeight || document.documentElement.clientHeight;
+        
+        if (rect.top <= windowHeight * 0.85) {
+          element.classList.add('visible');
+        }
+      });
+    };
+
+    // Initial check for elements in view
+    handleScroll();
+    
+    window.addEventListener('scroll', handleScroll);
+    return () => window.removeEventListener('scroll', handleScroll);
+  }, []);
+};
+
+export default useAnimateOnScroll;
